2-Django-Restframework-API/3-djangoDRF-debug-toolbar/requirements.txt
asgiref==3.6.0
Django==4.1.5
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
